New CLI command variation: "use either <item1>, <item2>,..." (also works with "eat", "drink", "hobo", and "ghost" commands).  This will use the first of the specified items that you already have in inventory, limited to the quantity in inventory.  Only if you have none of them will it attempt to make or buy the first item - in other words, it becomes equivalent to "use <item1>".
This was implemented specifically to handle the two combat% effects (Fresh Scent, Hippy Stench), each
of which is obtainable from two different items, and the default mood extenders for those effects have
been changed to use it.  I'm sure that users will come up with other creative uses for it.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6358 b29ace70-8910-0410-8dcc-aa2fc6433167