When in Ronin, items from storage can now be directly used from the consumption and creation lists within the Item Manager.  Pulled items will be shown in italic, and the number of pulls that will be used is shown in the queued ingredient lists.  To avoid clutter, consumables that you aren't high enough level to consume will not be considered pullable.
To enable this feature, you must first set a budget for the number of pulls
KoLmafia is allowed to use on your behalf - this is done in either the
General Storage or Equipment Storage lists (just above the display of the #
of pulls remaining).  Explicitly pulled items (from either Storage list, or
"pull" commands in the CLI) do not require a budget, but may reduce the
budget as it is limited to the number of pulls remaining.  The budget is
reset to zero for every session, to avoid surprises.

Use with caution!  If you've set a nonzero budget, mafia will happily use
those pulls to satisfy ANY item request that exceeds the amount creatable
from inventory!


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6441 b29ace70-8910-0410-8dcc-aa2fc6433167