Adds another scriptable behavior override: deciding whether to buy or create an item, which doesn't always have an obvious answer.  Set "buyScript" to the filename of an ASH script definining a function like this:   boolean main(item itm, int qty, int ingredientLevel, boolean defaultBuy) If it returns true, the item will be bought, otherwise created (possibly invoking further buy/create decisions for the ingredients).
Parameters itm & qty specify the item that is needed.

ingredientLevel indicates what's already available: 0 = none of the
ingredients, 1 = some of them, 2 = enough to create at least one of the
requested item.

defaultBuy indicates what KoLmafia would have done otherwise; returning this
value is the safest thing your script could do.  It will normally be true if
ingredientLevel is 0, false if 2.  The default at level 1 depends on the
item, and is subject to change - currently, anything made with star charts,
pixels, the Malus, or multi-use are bought, anything else is created.

Your script should generally NOT buy or create anything - those actions will
be performed based on the value it returns.  However, the script might want
to acquire the item by nonstandard means - for example, it could pull certain
items from the clan stash, while the general preference for satisfying
requests from the stash is turned off.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6466 b29ace70-8910-0410-8dcc-aa2fc6433167