Phase 1 of campground parsing - your housing, furnishings, and other campground equipment that isn't already detected is now collected in a list of items, available internally as KoLConstants.campground, and from ASH via get_campground(), which returns a mapping from items to integers (the integer can only be 1 for most items, but can be as high as 9 for pretty bouquets and fences).  There are a few special cases for campground items that don't correspond directly to an inventory item: having Really Good Feng Shui is represented by the item "Feng Shui for Big Dumb Idiots", having a pagoda is represented by "pagoda plans", and the ground (having no housing) is represented by a "big rock" for lack of anything more appropriate.
Modifiers are processed for all campground items, and new modifier types have
been created for the various aspects of recovering HP/MP via resting, but at
the moment the only items with any real effect are the Hobopolis bedding
items which give elemental resistance.  Also, the campground info is only
valid at the start of a session, it isn't yet being updated if items are
added.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6502 b29ace70-8910-0410-8dcc-aa2fc6433167