Remove carved wooden wheel when given choice adventure that consumes it, not when you take a choice.
Initial support for HiddenCityRequest: recognize and log visits to hidden city,
recognize when placing stone triangles or round objects consumes items.

This could ulitmately be expanded to allow detection and tracking of which
squares contain the various altars and the temple, to allow insertion of
round objects and stone triangles, and to allow autmated adventuring in
other squares.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6503 b29ace70-8910-0410-8dcc-aa2fc6433167