Adds a few CLI commands:
"try" - execute the following commands, but continue the script even if an
error occurs.

"else" - execute the following commands only if the previous "if", "while",
or "try" didn't execute theirs.

"elseif <condition>" - use this instead of "else; if <condition>" which
doesn't actually work.  Due to the unusual way the CLI defines block
statements, both "else" and the conditional command it goes with must be at
the outermost level; neither can be nested inside anything, thus a separate
"elseif" command is required.

"ash <statement>" - runs a single line of ASH code, and displays the results.
This is intended for quick testing of ASH built-in functions; for embedding
ASH code in a CLI script, the <inline-ash-script> mechanism remains a much
better choice.

Also fixed a crash that would occur if a return statement was used at the top
level of ASH code, outside of a function.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6510 b29ace70-8910-0410-8dcc-aa2fc6433167