Item creation panels in the Item Manager now create all selected items, not just the first.
Note the number of online players at login, so that the user has some idea of
how laggy this session is going to be.

Add Disembodied Hand to the list of useful Basement familiars; it can hold a
+stat% weapon for you.

Add "refresh campground" option to CLI, in case that's the one thing that
needs refreshing (it would otherwise require a "refresh all").

When refreshing, don't display the item being refreshed if that would be the
complete inventory list, or something equally large.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6597 b29ace70-8910-0410-8dcc-aa2fc6433167