"recover" command now recovers up to your autorecovery target if the current value is below it, and only then recovers one point at a time.  This makes no difference with the built-in recovery mechanism, which has a special case that accomplishes the same thing, but makes a lot more sense with a recoveryScript.
The HP & MP values in the relay browser are now clickable if they're less
than the maximum, rather than the autorecovery target.  This makes it easier
to recover a bit more when needed; healing up for a boss fight, or making
sure that you'll be able to cast Olfaction, for example.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6698 b29ace70-8910-0410-8dcc-aa2fc6433167