Add option to CLI "familiar" command: if you say, for example:     familiar naked baby gravy fairy (as opposed to simply "familiar baby gravy fairy"), KoLmafia will remove the familiar equipment after switching to the specified familiar.
This seems pretty useless except in one place: on the page where you are about
to step through the gash and ascend, you are offered a list of your familiars. 
If you select one from that list, KoLmafia will switch to it using this command,
so it will be as naked as you are and ready to enter the afterlife.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6717 b29ace70-8910-0410-8dcc-aa2fc6433167