Yendorian Finery no longer has a -combat effect.
Phase 2 of modifiers overhaul: Added "W" variable to modifier expression
language, representing the current familiar weight.

Modifiers are now processed for the curren familiar's race, prefixed by
"Fam:".  These modifiers can set the new values Volleyball, Sombrero,
Leprechaun, and Fairy, which are treated as weight multipliers for these
common familiar effects.  The values are assumed to be 1 if not otherwise
set, and the familiar is indicated as being of that type in familiars.txt, so
entries in modifiers.txt are only needed for familiars with nonstandard or
variable effectiveness.

Familiar type "item1" in familiars.txt has been removed: only the Hound Dog
was in that category, and it is now fully expressible via modifiers.

Modifiers are processed a second time for the current familiar equipment
item, prefixed by "FamEq:", thus allowing such items to have an effect that
doesn't apply when equpped in a different slot (specifically useful for the
Hatrack).  Any text in a Familiar Effect modifier of the general form
"<number> x Volley|Somb|Lep|Fairy" will be autoconverted into a FamEq
modifier, so that Hatrack hat effects don't have to be specified twice.

With this update:
* The Hound Dog and Exotic Parrot are now described entirely via the data
files, rather than special cases in the code.
* The Hobo Monkey, Wizard Action Figure, Llama Lama, and Mutant Rattlesnake
now have their proper effectiveness levels.  The Fire Ant and Cactus Bud can
be added once their exact effectiveness has been spaded.
* The Crimbo PRESSIE now switches between Sombrero and Volleyball behavior
depending on its equipment, although that doesn't do much good at the moment
since sombrero stat gains are still not being estimated.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6776 b29ace70-8910-0410-8dcc-aa2fc6433167