Use more accurate formulas for familiar meat & item drops.
Include intrinsic effects in the active effects list, which is specifically
needed to handle Temporary Blindness's item drop penalty.  That effect can
also be obtained as a non-intrinsic, and the two instances stack: therefore,
add a duplicate copy to the effects database with an "(intrinsic)" suffix,
and use the appropriate one when parsing the charpane.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6781 b29ace70-8910-0410-8dcc-aa2fc6433167