Track Hobopolis sewer tunnel explorations in the session tally.  No attempt is made to track dungeon reflooding or clan hopping - it's your responsibility to remember the number across sessions if it's important to you, or to zero it if you're going to participate in more than one dungeon instance during the same session.
Track total ML of monsters that the jam/rock band flyers have been used on,
in preference "flyeredML".  The total itself is not yet displayed anywhere,
however the amount you've done in the current session is shown in the tally. 
Not tested at all.

Don't bother casting Ode or giving an overdrink warning when drinking the
steel margarita.

Disable the Friars blessing buttons in Daily Deeds prior to level 6.

Disable HTTP timeouts when loading valhalla.php.  This may help make
ascension logging more reliable, as it's a really slow page load that cannot
be usefully retried.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6835 b29ace70-8910-0410-8dcc-aa2fc6433167