Custom outfits can now trigger actions when equipped, by including one of the following strings in their name:
F=<familiar> - change familiar
E=<item> - equip familiar item
M=<mood> - change mood
C=<command> - arbitrary CLI command

This should be especially useful in outfits for automated basementing, for
example "element f=parrot" to switch to your Exotic Parrot for elemental
tests.

Note that the commands are executed BEFORE the actual outfit change; this is
necessary in order for the charpane to show the changed familiar (or other
result) if the outfit was changed from the relay browser.

As a side-effect, putting on a custom outfit is now recorded in the session log.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6876 b29ace70-8910-0410-8dcc-aa2fc6433167