It is possible for inventory.php to both return a message and display our current equipment. Parse both, in that case.
When you insert a punchcard into an El Vibrato helmet, log it appropriately and
remove the punchcard from inventory.

Ditto for inserting a punchcard into an El Vibrato Megadrone. The five cards
that work in this capacity are now usable from the Item Manager.

(We don't currently remember the punchcard state of the helmet or familiar.)


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6889 b29ace70-8910-0410-8dcc-aa2fc6433167