Add Nardz energy beverage as a MP restorative option.
Add developer command "fakeitem <item>" that will add an arbitrary item to mafia's inventory.  This is useful for verifying that a new item appears in the right list, with correct consumption data, etc. without having to actually acquire one.  Actually using the fake item will fail miserably, of course.  Does not appear in the output of "help".


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6891 b29ace70-8910-0410-8dcc-aa2fc6433167