Phase 3 of special combat actions: adds automatic Olfaction/odor extractor and Spooky Putty sheet actions.  When enabled, these actions will prompt you for the conditions under which you want them to be used; they can also be enabled via CLI commands "olfact" and "putty".  As with the auto-ID actions, if you are using a CCS it must contain a "special action" line to indicate when you want the action to take place - perhaps just after Entangling Noodles, so the monster doesn't get a free hit on you.
The options for these commands are:
* nothing - shows current trigger setting.
* "none" - disables trigger.
* an item, or comma-separated list of items, optionally prefixed with "item"
for clarity - triggers on next monster encountered that is capable of
dropping all the specified items.
* "goals" - as above, but the list of items is taken from your current
conditions.  If no one monster is capable of dropping all the items,
triggering will be delayed until enough of the individual goals have been
satisfied for there to be such a monster.
* "monster" followed by a (partial) monster name - triggers on the first
monster encountered with a matching name.  The "monster" prefix is required
because many monster names are also unambiguous item names, due to their tiny
plastic or stuffed versions.
* Any trigger specification can be followed by "abort", to stop adventuring
when triggered even if there are unsatisfied goals.

If either of these actions are active, no autoattack will be automatically
set (unless it's Entangling Noodles), so you don't accidentally kill the
monster before tagging it.

Also: the "stop" button on the Adventure panel has been split into two -
"stop now" (with the previous behavior) and "stop after" (which cleanly stops
after the current adventure is processed, never breaking into combat).  This
is useful if you realize you've started adventuring with the wrong equipment
or buffs; you want to fix the problem as soon as possible, but having to
finish an interrupted combat in the mini-browser does you no good at all
(unless you were planning to use a free-runaway item).


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6921 b29ace70-8910-0410-8dcc-aa2fc6433167