Preliminary support choice adventure 309 (Barback): just like the equivalent Pastamancer deep sea choice, since you can only usefully use the first choice three times a day, track the number of seaodes (batter) you've obtained and annotate the choices with this information when you find the choice in the Relay Browser. We do not yet automatically choose to ignore the adventure if you've already obtained your daily maximum. Soon...
Certain things (like detecting progress through Hobopolis sewers) print a status
message - even if you are doing the actions in the Relay Browser. Doing those
actions from the GUI enables the display when you are done with the sequence,
but doing them from the Relay Browser leaves the display disabled. I identified
several such cases and changed them to re-enable the GUI, while I consider if
there is a general solution for the issue.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6979 b29ace70-8910-0410-8dcc-aa2fc6433167