When looking up an item in the NPC store database, use exact name matching only.  A profiling run showed that this was taking an absurdly high percentage of the time: refreshConcoctions() checks everything to see if it's available from an NPC store, and for the concoctions that aren't actual items (i.e. sushi), this was resulting in lengthy, futile attempts at fuzzy matching.
Rearranged the parsing of item drops from monsters.txt so that it tries
stripping off the drop rate number first.  In the vast majority of cases,
this avoids a useless fuzzy matching attempt.  (Also fixed a typo in
monsters.txt that was being covered up by the old parsing strategy.)

After these two changes, profiling does not show anything that seems to be
taking an unexpectedly large amount of time.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6983 b29ace70-8910-0410-8dcc-aa2fc6433167