Use cached Mall prices to more intelligently decide between buying the final product vs. buying missing ingredients.  Note that this still needs some tuning - in particular, if you are able to create one of the ingredients, it will be valued at autosell price (the same as if you had it in inventory already), not taking into account any NPC purchases you'd have to make to create it.
When creating items from the CLI, don't temporarily turn off
autoSatisfyWithNPCs unless the specified quantity is "*" or negative (in
which case this is necessary in order to calculate a meaningful "as many as
possible" value).  This will generally save a call to refreshConcoctions()
that is particularly slow because it actually changes the creatability of
many items.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@6990 b29ace70-8910-0410-8dcc-aa2fc6433167