Fix recipe for 17-alarm saucepan: we had it listed under 17-Alarm saucepan. That's consistent with the 5-Alarm saucepan, but it's not correct name.
When your Wild Hare gives you an extra rollover adventure, tally it in the
"extraRolloverAdventures" setting, which is (obviously) cleared at rollover.
The little pop-up on the sidepane showing modifiers adds this to the Rollover
Advs number from equipment, to give you the total number of extra adventures
you get at rollover.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7106 b29ace70-8910-0410-8dcc-aa2fc6433167