When KoLmafia is acting as a proxy for another browser - i.e., you are running in the Relay Browser - we accept requests from the browser and pass them on to KoL, and get the replies and pass them back to the browser.
Used to be no matter whether the browser issued a GET or a POST, we turned it
into a POST and sent it along.

Unfortunately, there are some pages on KoL - like the clan roster - which have
forms that specifically request a GET - and don't do the "right" thing if we
request the form via a POST.

We now recognize when the browser is asking for a GET or a POST and attempt to
Do The Right Thing in either case.

My testing finds no errors, so far - and the clan roster pages now work as
expected - but this is a big change. Don't hold back on reporting bugs! ;)


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7134 b29ace70-8910-0410-8dcc-aa2fc6433167