Add new quest items.
Fix boss warning and session logging for the Goblin King.

Add alias for the bogus URL on the bat hole map for the Boss Bat: there is a
trailing semicolon.

Log GET and POST requests from the browser in the debug log.

Efficiency change in Relay Browser: when you execute a charpane request (using
an up arrow to increase turns of an effect, executing a mood, restoring HP or
MP, etc.) do not do a charpane request between every request in the sequence and
then again at the end. Do it only at the end.

Custom javascript in basic.js now makes requests via GET, rather than POST. This
should prevent KoLmafia from adding unnecessary password hashes and otherwise
mucking with the fields.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7206 b29ace70-8910-0410-8dcc-aa2fc6433167