Do not use static EquipmentRequests when refreshing inventory. Doing so was half of the "Adventuring area not available" issue some people have reported:
- The session times out
- The player hits the "refresh inventory" button
- The requests get redirected to login.php?notloggedin=1
- Since the requests are configured to automatically follow redirects, the URL
in the request is changed to the not-logged-in page
- That page does not list your inventory, so KoLmafia can no longer read your
inventory - static requests! - until you close down and restart the program.

Revision 7243 fixed the failure to timein equipment/inventory requests. But,
for additional safety, this revision eliminates the static requests.

Robbing the Deluxe Mr. Klaw was still trying 4 times...

Attempt to eliminate unnecessary login server ping pong: when we are directed to
use a different server, obey it, as before, but also use the sessionid we were
given, rather than the abandoned one that forced the redirect the first time.

It's "The Remains of a Jilted Mistress" - complete with "The".

Fix changed enchantment description for bottle of moontan lotion.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7280 b29ace70-8910-0410-8dcc-aa2fc6433167