Add an EQUIP_MATCH filter type to the ItemFinder, and use it for CLI "equip" commands.  "equip damned" no longer considers the Children's Meal of the Damned to be an option!
If exactly two items match an ambiguous name, and one is a substring of the
other, assume the shorter name, on the grounds that the user would have typed
part of the unique portion of the longer name if that's the item they
actually intended.  This should help item matching in two specific cases:

1. Non-clockwork in-a-boxes and their components can now be abbreviated. 
Previously, they always had to be typed in full to disambiguate them from
their clockwork equivalents.

2. DoD potions can now be referred to by flavor.  Previously that would
produce a silly ambiguity error, like "dark potion" vs. "dark potion of
blessing".


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7294 b29ace70-8910-0410-8dcc-aa2fc6433167