Per-user preferences with names starting with an underscore are reset to their default values on the first login of each day, without needing to individually hard-code them in KoLmafia.resetCounters().  This ability can also be used by user-defined preferences (in which case the default value is an empty string), which may eliminate the need for scripts to have logic to determine whether or not this is the first login.
Added five new preferences using this feature: _navelRunaways and
_banderRunaways, tracking free runaway uses; also _gongDrops, _absintheDrops,
and _astralDrops, tracking special familiar drops.  All of these values are
displayed in Daily Deeds.

If the chance of a runaway being free is non-zero, annotate the old-style Run
Away button with this info.  Unfortunately, there's no corresponding display
yet when using the CAB, as it's not so easy for mafia to modify.

CCS "run away" action can now contain a number; running away will only be
attempted if the percentage chance of it being free is at least as great as
that number.  Otherwise, the action will be skipped, and the CCS will
continue with the next line.  Examples: "run away 100%" will only run if it's
guaranteed to be free.  "run away 50%", with a navel ring equipped, will
attempt running on the first 8 combats in a day (the 3 guaranteed free
runaways, plus 5 more attempts with increasing chance of costing a turn).

Very little of this has been tested, as I lack all of the relevant familiars.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7302 b29ace70-8910-0410-8dcc-aa2fc6433167