Adds seal summoning & hot tub soak counts to Daily Deeds.
Modifier names (as used by ASH numeric_modifier(), etc.) are now
case-insensitive.

If auto-adventuring in the Basement changes your outfit, clear all outfit
checkpoints.  The resulting item-at-a-time restore is painfully slow, and
completely useless - to continue, you're likely to want some variation of the
current outfit, NOT the one you happened to start with.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7313 b29ace70-8910-0410-8dcc-aa2fc6433167