Use commas in the prices shown by the "cheapest" & "expensive" commands.
When inquiring about effect modifiers that depend on the remaining turns of
the effect ('T' value in the modifier expression language), for an effect
that the player doesn't currently have, pretend that there is one turn of the
effect.  This allows anything that scans the modifiers looking for particular
boosts (such as the Basement helper) to realize that such effects actually do
something.

Added some missing effects, and the Slime Hates It modifier.

Added an override mechanism for modifiers: setting preference _userMods to a
modifier specification will include them in the player's modifiers whenever
they are recalculated.  This can be used to get a correct display of a
modifier if there's some source that mafia doesn't know about yet, or doesn't
track.  For example, if you know that there have been five gall bladder
squeezes in your current Slime Tube run, you can adjust the displayed ML
value via:
  set _userMods = Monster Level: +100
Or better yet (since it would only apply in the Slime Tube):
  set _userMods = Monster Level: [100*zone(slime)]

See src/data/modifiers.txt for examples of how modifier specifications are
written - note that they are case-sensitive, and require precise formatting. 
Use "modref _userMods" to see exactly how your override was parsed.

Since this preference starts with an underscore, it will be cleared on the
first login of each day.  This prevents forgotten-about changes from
persisting forever.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7368 b29ace70-8910-0410-8dcc-aa2fc6433167