Fix problem introduced in r7382 with the use of combat-causing items (such as drum machines) from the relay browser, with the Combat Action Bar enabled. The symptom was that the first round of combat overlaid the mainpane, rather than replacing it.
Don't request a charpane update in response to chat notification that you've
received a buff, if mafia is currently processing a command.  The
asynchronous update might not get processed in the proper order relative to
the updates performed by the command, resulting in an incorrect adventure
count or possibly other problems.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7386 b29ace70-8910-0410-8dcc-aa2fc6433167