Initial release of the Modifier Maximizer, a tool for suggesting ways of boosting an arbitrary numeric modifier (or a weighted sum of multiple modifiers).  It is available from the Tools menu, can be added to your startup tabs or windows, and can also be invoked via the CLI "maximize" command (although that's not very useful yet).  It is currently mostly working, for effect-based boosts only; equipment suggestions are planned in the future.  There is a Help button in the interface for more info.
The core change needed to support the Maximizer is the ability to calculate
modifiers on a speculative state, rather than the actual player state.  This
is still a work in progress: speculating about equipment changes is made
difficult by the fact that all the equipment accessors are static methods,
which Java provides no way of redirecting.

It is now possible to have string modifiers at the character level.  There
are three current instances of this:
* "Equalize" holds the value from the first equalizer potion seen in the
active effects list, which seems to match KoL's behavior in the unusual case
of having more than one of them active.  This allows stat calculations to
take equalizers into account without having to do additional searches through
the effects list.
* "Intrinsic Effect" is set to a tab-delimited list of all such values from
current equipment items.
* "Outfit" is a new modifier, set to the name of the standard outfit being
worn, if any.  This is determined in the process of calculating modifiers, so
it might as well be made available without having to recalculate it.

"modref" has been updated to display character-level string modifiers, and
ASH function string_modifier(string) has been added to allow script access to
them.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7460 b29ace70-8910-0410-8dcc-aa2fc6433167