The plural of "tasty louse" is now "tasty lice".
Weapons equipped on a Disembodied Hand cannot grant Slime Hates It or
Brimstone bonuses.

Added a new class of modifiers - bitmaps, in which each bit represents a
distinct source of something.  This is a generalization of my earlier code
supporting Clownosity, which is exactly what is needed for Brimstone as well
- both only count distinct pieces of equipment.  In addition to those two,
two other bitmap modifiers currently exist:
* "Synergetic" indicates items that grant bonuses in conjunction with one or
more otherwise unrelated items.  This will allow such synergies to be
efficiently detected without having to be hard-coded in the program, although
the actual detection has not been implemented yet.
* All boolean modifiers now exist as bits within a single bitmap, rather than
an array of individual booleans.  This allows an item's booleans to be merged
into the character booleans with a single OR, rather than testing them one by
one.  This will become more relevant in the near future, as I'm planning to
add at least 18 new boolean modifiers in support of the Modifier Maximizer.

Bitmap modifiers can be queried from ASH using any of the existing forms of
numeric_modifier() - note that the value will be the count of bits, rather
than the raw bitmap value itself.  "modref" has been updated to display
bitmap values, but "modtrace" does not yet support them.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7461 b29ace70-8910-0410-8dcc-aa2fc6433167