Fixes to MMG support:
When you place a bet, KoL responds by redirecting you to bet.php. Only after you
follow the redirect do you learn whether you ducceeded in placing the bet and
have the opportunity to find out what the bet id of your new bet was. This extra
request/response creates two race conditions:

1) Between the time KoL sends you the initial response with redirect and the
time it receives your redirected request, somebody can take your bet. If so, KoL
will not include your new bet on your list of pending bets and we cannot learn
the bet ID.

2) Between the time KoL makes the bet page which includes your bet and the time
you receive it, somebody might take your bet, generating a chat event - which
you can receive before the bet response page

Handle both of these by saving the amount we bet before sending off the request.

If we can't find a new bet on the response (race #1), make a generate a dummy
bet with a negative bet ID, to be matched by the chat event.

If we receive a chat event and can't match it with a pending bet, see if we
have an outstanding but not-yet-registered bet for that amount, and save the
chat event to be matched later.

3) If you fail to place a bet (not enough Meat, no casino pass), notice this
and don't assume that just because the bet does not appear in your pending bet
list it must have been taken already.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7505 b29ace70-8910-0410-8dcc-aa2fc6433167