ASH foreach loops can now contain one more variable in the list than the target map has indices; the additional variable, if present, gets assigned the current value from the map (saving you the trouble of explicitly looking it up inside the loop).  In the case of multidimensional maps, this is more than just a convenience, it's a definite performance boost: the internal code can directly look up the value in the final slice of the map, whereas a script would have to perform the entire chain of lookups starting from the top.
This change makes it useful to iterate over arbitrary aggregate values (such
as the return value of a function), not just aggregate variables, so that
change was made as well.  Note that this is only useful when the extra value
variable is present, as otherwise you'd have no access to the actual values
in the map.  For example, you can now say "foreach itm, count in
get_ingredients(someItem) {...}" - you don't have to assign the map to a
temporary variable and then iterate over that.

This introduces a small backwards compatibility problem: a loop with a body
that's a parenthesized expression (rather than the usual block or simple
statement) would now be parsed as if the target was a function call, with the
body being interpreted as the parameter list.  This seems like an extremely
uncommon usage, which can be easily fixed by putting curly braces around the
loop body (or parentheses around the target).


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7514 b29ace70-8910-0410-8dcc-aa2fc6433167