Add support in WumpusManager for Wumpinator - For each room, track how many other rooms report each potential hazard - Add "wumpus status" command which displays that data - Calculate Wumpinator code - Add "wumpus code" command which displays the code - Add menu option to call Wumpinator with current code - Whenever you are given a choice in the Wumpus Cave, add a link to "View in   Wumpinator", which invokes the Wumpinator with the current code.
KoLmafia's deductions remain the same as ever - and are easily seen to be
inferior to the Wumpinator. I still am unable to reproduce an INCORRECT
deduction, but the WUmpinator deduces things we can't deduce, yet.

I also have the code to actually link to the Wumpinator's generated graphic,
giventhe code, but I am missing one additional input parameter to make it
actually work. Soon, I hope...


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7562 b29ace70-8910-0410-8dcc-aa2fc6433167