When calculating creatable amounts, handle ingredients with nonstandard acquisition methods (Hermit items, Tr4pz0r furs) a bit better.  The code to handle them was sometimes triggering on other items, causing their available quantity to be exaggerated.
Fixed problem that was causing mini-browser buttons to be unclickable if they
contained any HTML entities, such as "Attack with your Coily&trade;".  There
was some strange code in RequestEditorKit that was determining the
RequestFrame that contains a form submit button by enumerating all Frames,
looking for one that contained the button name (which failed in this case
because the entity was mangled) - rather than simply asking the button view
for its parent.  The problem with mangled entities remains, something is
eating the final semicolon from them (except for nbsp, apparently), but I
haven't been able to determine what it is.

Fixed another class of broken JickHTML so that the Haunted Wine Cellar
appears properly in the mini-browser.

The Wine Cellar decorator now indicates the correct order of dusty bottles as
1st/2nd/3rd, rather than 1/2/3 which could easily be misunderstood as being
inventory counts.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7588 b29ace70-8910-0410-8dcc-aa2fc6433167