Fixed level requirement for roc feathers.
Fixed logic of use link for the dinghy plans when you don't have the dingy
planks already: it would send you to the Hermit only if you didn't have any
worthless items, and would send you to the sewer if you did.

Fixed handling of multiple counters expiring on the same turn - it wasn't
always ignoring counters that had expired in the past.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7719 b29ace70-8910-0410-8dcc-aa2fc6433167