InventoryManager.getAccessibleCount() now counts items equipped by all your other familiars, in addition to your current equipment.  (This doesn't affect ASH available_amount(), but perhaps it should.)
Modifier Maximizer now supports "switch <fam>" to consider switching to a
different familiar.

Maximizer will not consider any item you don't actually have unless it has an
actual benefit - that protection previously only applied to items that would
be bought from the Mall, not NPC purchases/creation/etc.

Maximizing "adv" now implies "-tie".  In the unlikely event that someone
wants to use the tiebreaker in this case, spelling the modifier in full as
"adventures" bypasses this.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@7858 b29ace70-8910-0410-8dcc-aa2fc6433167