If something goes wrong with updating the text of the mini-browser sidepane (I've seen some random inexplicable NPEs here recently), don't let the exception propagate - that can kill one of the relay agent threads, resulting in a lost pageload.
Fixed problem with slime vials becoming uncreatable for the rest of the
session once identified, because the corresponding concoction was being
looked up via name (which has changed) instead of ID.

The "bang" and "vials" commands now show quantities - on hand, in the closet,
and creatable.  This should be useful to Saucerors to find a vial that they
can make, but haven't identified yet.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@8225 b29ace70-8910-0410-8dcc-aa2fc6433167