ShowDataCommand now tries to match its filter against a version of each data item with HTML tags stripped out, so that commands like "inv elven limbos" work as expected.
Clear the error state before launching a .ash relay script, so they can be
used immediately after an error of some sort without having to manually clear
it.  Note that this does NOT apply to actual relay overrides, as it appears
that leaving the error state alone is an unfortunate necessity in that case. 
In particular, if an override of charpane.php existed, it would be constantly
clearing the error state, which would at least occasionally lose an actual
error that occurred during between-battle processing.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@8335 b29ace70-8910-0410-8dcc-aa2fc6433167