Improve handling of malformed combat macro results.  Note that there is currently a server-side bug causing "skill" commands with an ID (rather than a skill name) to be misreported: this revision will avoid a blank relay browser page, but the combat won't actually be logged correctly.
Added an error trap in GenericRequest that should make it less likely for
botched parsing to result in a blank relay browser page.

Added basic syntax checker for combat macros, available on both the macro
management page, and on the fight page itself if you've set macroLens to
reveal the on-the-fly macro field.  The check is automatically performed
after 2 seconds of inactivity.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@8406 b29ace70-8910-0410-8dcc-aa2fc6433167