Initial release of CCS macrofication: there are still some differences from round-by-round combat (for example, in-combat MP restoration isn't working yet), but combats can now generally be converted into a KoL macro and resolved in a single page load, as long as they contain only eligible actions.  Ineligible actions currently include "consult" (utterly impossible to macrofy), "jiggle" (no macro equivalent yet), and "delevel" (tricky to duplicate in a macro).
There are a couple of inherent differences in macro combat actions:
* After a CLEESH, or other monster transformation, combat continues with the
same CCS section, rather than possibly jumping to a section that applies to
the new monster.  If "skill CLEESH" is the final action, an attack loop is
added to finish off the amphibian.
* CCS "abort after" will always abort automation afterwards if the section
contains this command, rather than only aborting if combat reached that line
of the section.

Setting pref "macroDebug" to true will display the generated macro for each
combat in the gCLI.  This is the most human-readable format of the macro; a
more optimized version, with uncalled and empty subroutines stripped out,
appears in the debug log.  (And the version of the macro actually sent to the
server further has comments and leading whitespace stripped out.)

The CCS editor will now accept "use" as a synonym for "item", for
compatibility with KoL macros.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@8416 b29ace70-8910-0410-8dcc-aa2fc6433167