Various improvements to disco combos:
In the DB combat helper, the combo names are now buttons which will submit
the entire combo as a macro.  The buttons will be disabled if you don't have
enough MP for the entire sequence, or there's already a possible combo in
progress.

Added "Random Rave" combo, which only exists while you know all three rave
skills, but not all six rave combos - it will be an arbitrary sequence of the
three skills that doesn't correspond to a combo you know.

Added "combo" CCS action to perform a combo by name - "combo rave nirvana",
for example.  It works in both macrofied and round-by-round combat.  MP
restoration (if enabled) is performed once, for the total amount needed by
the combo, and likewise all every-round actions (like automatic antidote use)
are suppressed during the combo, to avoid breaking it.  The action will be
skipped without error if you're not a DB, don't know some of the individual
skills, don't know the combo itself in the case of rave combos, or have
insufficient MP for the entire combo and don't have in-combat restoration
enabled.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@8428 b29ace70-8910-0410-8dcc-aa2fc6433167