Allow a [global prefix] CCS section to contribute to a generated macro, even if a simple battle action is being used.
Bricks claim to be a gift item, but can't be given to other players via kmail
- they can only be thrown.  Mark them as a display-only item, instead.

Update the filter on the built-in kmail attachment list, so that untradeable
items are hidden from the start.

If an opera mask is equipped, restore MP before any skill use, rather than
using up your existing MP before any restoration.  This maximizes the benefit
from the mask.

When equipment is removed in combat by Zombo, update the equipment and
inventory lists appropriately.  Not tested.

Fun Bluegrass Fusion delevels by 20.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@8490 b29ace70-8910-0410-8dcc-aa2fc6433167