records.php is now exempt from result processing, as there is at least one person out there with a familiar with a name starting with "You acquire".
The Chatterboxing choiceadv now has additional options to use a valuable
trinket for banishment if you have one, but otherwise get a boost to a
specified stat (or your mainstat).

Really thick bandages are now properly flagged as a restorative.

The player damage regex now accepts "like this, it's bound" as a valid
suffix, needed for one of the Disco Banjo hit messages.

Entangling Noodles can be used multiple times against the Spaghetti Demon.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@8501 b29ace70-8910-0410-8dcc-aa2fc6433167