When automating a Fernswarthy's Basement stat or mpdrain test, and the relevant outfit is either nonexistent or insufficient to pass the test, the Modifier Maximizer is now invoked automatically.  (The same isn't done for element and hpdrain tests, because the Maximizer cannot currently express the optimum balance between resistance and HP needed for those tests.)
Added some hidden prefs to fine-tune the mana burning process:
"skillBurnXXXX" (where XXXX is the skill number) sets the priority of this
skill: -100 or lower disables it for extension, numbers from -99 to 0 enable
it for extension up to (N+100)% of the burn limit, normally 1000+turns
remaining.  The value is offset like this so that missing prefs read as 100%
by default.  All previously hard-coded unextendable buffs (like combat rate
modifiers) now simply have a default skillBurnX of -100, so they can be
overridden if desired.  Values greater than 0 are reserved for possible
future use to indicate that the skill should be auto-cast if there's nothing
else to spend burned MP on.
"lastChanceThreshold" is amount of unburned MP at which alternatives to
extending existing buffs will be considered, 100 by default.  There is
currently one alternative implemented: "lastChanceBurn" will be executed as a
CLI command, with any instance of "#" replaced by the amount of MP that needs
to be used up somehow.  This could be used to invoke a script that records AT
buffs, for example, or could simply be an "echo" or "abort" command that lets
you know you need to find something else to spend MP on.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@8871 b29ace70-8910-0410-8dcc-aa2fc6433167