Add new chat command to KoLmafia: /trivia
Typing /trivia into chat re-introduces the old KoLmafia feature of combining all PMs into one tab and (in case you aren't actually using KoLmafia chat, and also to make it easier to bulk-send items or buffs to winners) opens up KoLmafia's pop up a contact list window specific to trivia games.  All PMs will be redirected to the one tab (which will be [trivia#] where # is an internal ID used to track each use of the /trivia command) and all people who send you a message will be listed in the contact list window in order of PM receipt.

In order to start a new game, simply typing /trivia again.  In order to disable this feature and return to normal "each PM opens a new tab" behavior, type in /endtrivia or /stoptrivia.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@8965 b29ace70-8910-0410-8dcc-aa2fc6433167