Instead of loading request sync, load up the relay browser with the page by cheating and loading up game.php and intercepting the subsequent request for main.php.
In this evil plan, the request for main.php results in a redirect for the page that generated the error (though without any URL string data).  This request is also intercepted and KoLmafia feeds the browser the response text that would have been shown in the request sync, except it gives KoLmafia a chance to at least try to reformat the response for the relay browser the way it usually does for other relay requests so that the user isn't left with the vanilla interface.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@9031 b29ace70-8910-0410-8dcc-aa2fc6433167