Update chat_private to not require that the user you're sending the message to be a clan member to allow for larger clans to take advantage of the feature.  Of course, you could technically flood someone, we'll hope that people act responsibly.
Add chat_notify command which prints the first parameter in the color specified by the second parameter.  Note that internal chat has a limited number of colors that it accepts, though relay browser chat will theoretically accept anything.  This feature allows personal scripts to notify the user's chat rather than only print messages to the gCLI where the user might not be looking.

Update internal message sending to also broadcast messages to relay chat by creating a fake chat history entry to be re-sent when sending received messages.  This implementation is not sufficient to re-broadcast messages sent from the relay browser to other relay tabs (we're going to assume that it's more trouble than it's worth to give each browser an ID and there's no other way to really do it reliably).

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@9043 b29ace70-8910-0410-8dcc-aa2fc6433167