Start of ChoiceManager refactoring, initially focused on collecting all the data and code for each choice adventure in a single place, rather than spread over 3 arrays and 11 functions.  I have only refactored the static data so far, but didn't want to hold up release of these other changes.
Added "go to goal" button for Dungeon Fist!, giving you a consistent 30 Game
Grid tickets.  This can only be done from the initial page of the game, as
it's just a playback of a recorded sequence of moves, with no ability to pick
up in the middle of a game.

There are 122 moves in the sequence, so the existing choice-goal mechanism
wouldn't work very well - the relay browser would almost certainly time out
the request before the sequence was complete.  Therefore, any command
submitted via /KoLmafia/specialCommand that takes more than half a second to
execute will now produce a status page that auto-refreshes every second until
the command is complete.

Recovery, mood maintenance, and other between-battle actions are no longer
attempted between the steps of a chained choice adventure (and hopefully also
multi-form battles like Ed, but I was unable to test that).  This was needed
for Dungeon Fist! because it replaces your normal HP readout on the charpane
with the in-game HP, which would otherwise be likely to falsely trigger
recovery.

The main item-usage commands are now blocked when a choice adventure or fight
is in progress, as they would otherwise fail in ways that we don't always
detect, resulting in inventory count discrepancies.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@9999 b29ace70-8910-0410-8dcc-aa2fc6433167