Remove certain trendiness checks, since items already in your inventory can be used regardless of their trendiness.
For trendy characters, do not count untrendy items in storage in InventoryManager.getAccessibleCount() and do not try to acquire untrendy items from the stash, the storage or the mall in InventoryManager.retrieveItem().

In the Maximizer, check if clan items are trendy when needed. Also check if you can equip suggested familiars.

In the Daily Deeds, check if you can equip familiars.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@10012 b29ace70-8910-0410-8dcc-aa2fc6433167