"N pirate insult" and "N arena flyer ml" are now valid goals.
The vast majority of AdventureResult instances now have their count as a
field, rather than a 1-element array - the memory and time savings are small,
but get multiplied by the many thousands of these objects that exist.  The
few instances that actually need an array with multiple values (3 or 4 at any
given time) are now handled by a subclass.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@10016 b29ace70-8910-0410-8dcc-aa2fc6433167