Various fixes to possible issues found using FindBugs (http://findbugs.sourceforge.net). None of these changes effect existing functionality, they are optimizations and may prevent possible exceptions in edge cases.
AdventureResult.java, chat/ChatParser.java, DwarfFactoryRequest.java:
Replace string concatenation in loops with a StringBuffer as the complier cannot optimize adding to a string when the concatenation occurs in a loop. While the gains in these particular loops may be tiny this is a good practice to follow.

persistence/HolidayDatabase.java:
Remove gameHoliday == null check in loop conditional, this check is unnecessary as the code returns both before the loop if gameHoliday is null and inside the loop if gameHoliday is not null.

request/FightRequest.java:
In the slim chance src returns null here, accessing image would cause a NullReferenceException. The only way I could foresee this occurring is if there were lag problems and the page doesn't load fully.

request/UneffectRequest.java
If effect is null, a NullReferenceException would occur when calling Kolmafia.updateDisplay. This was most likely supposed to use this.Effect, which should return the same effect name. This case is seems possible but extremely rare as calling 'uneffect' typically seems to return much earlier in the code path.

swingui/CakeArenaFrame.java:
Prevent a possible NullReferenceException in the event skill would be passed into the OpponentListener constructor when it is null.

textui/command/ShopCommand.java:
Move the call to getItemId() after the is null check so a NullReferenceException doesn't occur.

textui/Parser.java:
In this case result is guaranteed to be null, so pass in null instead of result to make the passed in value more apparent.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@10403 b29ace70-8910-0410-8dcc-aa2fc6433167