The ASH profiler can now get timestamps in nanoseconds rather than milliseconds, thanks to Java 1.5 (although you may not get any better than microsecond accuracy, depending on your platform).
Nowhere in KoLmafia are floating-point values in scientific notation allowed
- not in ASH numeric literals, not in modifier expressions, and not via
to_float().  It therefore seems inappropriate that ASH would ever produce
scientific notation when stringifying a float.  Of the two possible solutions
to allowing all floats to survive a round trip to strings, I chose the vastly
easier one of disallowing scientific notation on the output side. 
Unfortunately, Java (even 1.5) has no float formatting methods that can be
told not to ever use scientific notation, yet support the desirable feature
of "output only as many decimal places as needed to reproduce the value". 
They all require an explicit precision - which has to be at least 45 digits
to distinguish the float values closest to zero.  So, for now, stringified
floats are going to be a bit on the ugly side, although not in any way
inaccurate (there are known solutions to this problem, but they're either
slow or complicated).  Note that this approach is workable only as long as
we're using Java single-precision floats; a double in non-scientific notation
can require over 300 digits to accurately represent.


git-svn-id: http://svn.code.sf.net/p/kolmafia/code@11042 b29ace70-8910-0410-8dcc-aa2fc6433167