In headless mode, process commands asynchronously from their input.
Entering "--" on a line by itself aborts all retrieved lines up until that point and works like hitting the escape key in GUI mode.

You can put "--" into a script and get to effectively achieve a do-nothing but abort script (not terribly useful), or you can type it into the CLI in headless mode to get KoLmafia to stop execution.  The latter is the intent of this.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@11097 b29ace70-8910-0410-8dcc-aa2fc6433167