maximizer MRU is now opt-out.
MRU places your last few entries (default: 5) in the top of the maximizer dropdown.  Size of this list can be controlled by setting maximizerMRUSize; setting it to 0 will turn off this feature.  Default maximizer entries will remain at the bottom of the dropdown under a JSeparator (AKA: a line).

I know that it is perhaps confusing for maximizerMRULength to turn into maximizerMRUSize for no apparent reason, but it was the simplest way to reset that pref to default for everyone.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@12105 b29ace70-8910-0410-8dcc-aa2fc6433167