add preference simpleSvnUpdate (global, default true)
if true, simply checks if the root of your working copy is at HEAD, skipping it if so.  Speeds up the average update operation by 75% (and is probably nicer on servers).  Advanced users who have done something like a non-root "svn switch" or are using svn externals should know whether they want to turn this off.

Again the ticky box goes at the bottom of general preferences, it is soon time for these to be moved to their own page.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@12181 b29ace70-8910-0410-8dcc-aa2fc6433167