new ash svn functions:
boolean svn_exists( String projectname ) - returns true if a valid working copy named projectname exists in the svn/ folder

boolean svn_at_head( String projectname ) - returns true if projectname exists, is a valid WC, and is currently at the same revision # as the repository.  This will hit the repo server, so be nice.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@12198 b29ace70-8910-0410-8dcc-aa2fc6433167