Address issues with improper setting of s'more fullness.
The Item Manager sorts based on Concoctions that have their sizes
(fullness, inebriety, spleen) populated when the program is initialized,
presumably as a performance optimization. Because other variable
consumables (notably, astral consumables) have fixed size but variable
adventure gain, this was a nonissue in the past.

The solution here, therefore, is to re-evaluate the size of the s'more
Concoction (which in fact does vary) when setting its yields.

It is interesting to note that stat gains were not subject to these issues,
since those were re-evaluated in the Concoction comparator. Along the same
lines as that initial performance optimization, it may be sensible to also
store this information in the Concoction, and instead put the call to
objectpool.Concoction.setConsumptionData() in
persistence.ConsumablesDatabase.setConsumptionData(). I will not do that at
this time.

Miscellaneous changes that seem pointless are a result of running
M-x delete-trailing-whitespace on ConsumablesDatabase.java.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@16097 b29ace70-8910-0410-8dcc-aa2fc6433167