Migrate away from JList.getSelectedValues().
This method has been deprecated since Java 7. The recommended
replacement is to use getSelectedValuesList(), which unsurprisingly
returns a List<Object> (as opposed to an Object[]), so I've also
additionally gone through and added a .toArray() call to each one to
limit impact to the call sites.

There's potentially some broader cleanup to use this properly and
actually take advantage of these generics, but that can happen on a
later date (when I tackle -Xlint:unchecked).

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20832 b29ace70-8910-0410-8dcc-aa2fc6433167