Add Line, Token to Parser.java.
This is (hopefully) the biggest refactor needed as part of fredg1's
Language Server work.

We recognize this is a big change to unpack.

Some notes:

- We read the entirety of the InputStream in the Parser() constructor,
  parsing the input file into Line objects. We perform the parsing of
  Lines into Tokens throughout Parser.parse(), which continues to
  enforce many things ranging from variable typing to `since` version
  checking.

- For a non-null Token, the class's constructor guarantees that its
  String content is additionally non-null.

- Parser.currentToken() previously returned a non-null String. We now
  amend that contract such that it now returns a non-null Token, and
  as above, its content is also guaranteed to be non-null.

- Parser.nextToken() definitely does not provide these same guarantees
  (and has not in the past). However, in a number of cases, we
  observed that by making slight adjustments to when we read tokens,
  we can replace a number of calls to nextToken() ... readToken() with
  readToken() ... currentToken() without changing any underlying
  parsing behavior, while also improving safety at the same time..

Many of the scattered changes relate to either modifying existing
method signatures to take a Token instead of a String, or using
Token.content where we previously directly read a String token.

We've added a number of tests to codify what we expect as the current
behavior, and updated ParserTest to specify the expected tokens
resulting from a successful parse. It's often easier to reason about
behavior with concrete examples for various edge cases, rather than
from reading thousands of lines of code (regardless of quality of code
comments / documentation).

There are still some more changes coming down the pipeline to improve
parse consistency, most notably as it pertains to tokenization of
typed constants (both singular and plural). We expect that future
changes will be more manageable; this change was especially
challenging to split further because of how central parsing logic is
to ASH scripting, and how fundamental Parser.currentToken() is.

For future work: there are some segments of code that are rarely (if ever)
exercised (e.g. parsing "123." is unreachable because we first try and
fail to parse it as a record). But, related to this effort and adding
various test cases, we now know about these, so we can either delete
or fix them in the future.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20835 b29ace70-8910-0410-8dcc-aa2fc6433167