Move the "Missing index token" check later.
Also courtesy of fredg1, also in the interest of creating a language
server for ASH.

This puts it into the same if / else if / else chain we've set up
earlier, toward the same goal outlined in the previous commit. One of
the error messages for one of the test cases changes, arguably for the
better.

(If you specify "int[, ...] you get "Missing index token" rather than
"Invalid type name ','".)

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20869 b29ace70-8910-0410-8dcc-aa2fc6433167