Store info about outift Trick-or-Treat treats in mafia database and make accessible via `float [item] outfit_treats( string outfit )`
For now, outfits that drop an item/items randomly from a set will return that they drop nothing at all. The plan eventually would be for the `float` value in the response map to be a probability of that item dropping. This of course is a bit broken when the outfit can drop more than one of a set, but it'll have to do.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20872 b29ace70-8910-0410-8dcc-aa2fc6433167