Address gradle issues raised by MCroft.
lib/ shouldn't be part of the main sources; that was a hack to make
sure those files would be compiled. It turns out that we can specify
file dependencies built by a different sourceset, so do that
instead. This also simplifies some Clover configuration, since we now
no longer need to specify additionalSourceSets. (That in turn may
unblock proper incremental builds; I'll have to look into that again.)

Source files are ignored in the resources source sets, but Intellij
doesn't care and will complain anyways. So, add a manual exclusion for
all .java files in the resources directory.

Lastly, IntelliJ doesn't respect the clover source / target versions,
and we probably should be adding those for non-test builds
too. (Similarly, we should make sure that the source code is always
treated as UTF-8, in the name of portability.)

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20896 b29ace70-8910-0410-8dcc-aa2fc6433167