Read initial tokens in various parseFoo methods.
Courtesy of fredg1, as another entry in the effort to bring language
server support for ASH.

This combines two separate but logically similar patches, which affect
parseAggregateLiteral, parseNewRecord, and parseTypedConstant. In
these cases, we want to associate the initial token with the rest of
the language construct.

(We probably could make parseAggregateLiteral closer to these other
two methods, checking for the leading { and returning null if it
doesn't exist, but that's not so important to fix right now.)

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20902 b29ace70-8910-0410-8dcc-aa2fc6433167