Eliminate Scope(ParseTreeNode, BasicScope).
Courtesy of fredg1, as yet another patch for language server support for ASH.

We can accomplish the same thing by adding the command to the scope in the
process of parsing it. This patch further combines some of the logic so we don't
need to create a new Scope in several different locations.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20905 b29ace70-8910-0410-8dcc-aa2fc6433167