Stop discarding InputStream in favor of File.
Courtesy of fredg1.

This will be relevant as part of ASH language server support, since we may want
to provide both a File (location) and an InputStream (modified contents).

This patch also fixes an issue with circular imports -- if the base file is
imported by one of its dependencies (including by itself), then it should not
re-execute all the top-level code (which could lead to function redefinitions,
leading to unhelpful error messages, or worse).

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20907 b29ace70-8910-0410-8dcc-aa2fc6433167