Rearrange more logic in Parser.java.
Courtesy of fredg1, as yet another installment moving us toward
language server support in ASH.

This shuffles around some more if-statements so when we remove these
early returns (esp. in the context of throwing exceptions), we can
continue parsing.

This patch focuses on cases where the interpreter parses tokens in a
loop, e.g. when parsing parameters for a function prototype, or
statements in a block.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20908 b29ace70-8910-0410-8dcc-aa2fc6433167