Remove lib/org/jdesktop, merge {lib,src}/jar.
This change removes the manually-included swingx sources in favor of
bundling the jar directly, in part because there are a number of
deprecation warnings in whichever version we were using.

This leads to a slight increase in the jar size (a few hundred KB by
my measurement), since we're now pulling in everything from swingx,
rather than just what we need.

There are no longer any jars that are required just for lib/ sources
(swingx is now required for both), and the two remaining jars were
actually dependencies of existing jars in src/jar.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20918 b29ace70-8910-0410-8dcc-aa2fc6433167