Generalize TrayIcon support.
As of Java 6, cross-platform tray icon support is part of the standard
library. Thus, we no longer need this special library that only worked
for 32-bit Windows anyways.

I tried this briefly on a Windows machine, and it's a little weird
(especially around clicking on it when not logged in), but it seems to
work.

Note that in order to use this, useSystemTrayIcon must be true.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20923 b29ace70-8910-0410-8dcc-aa2fc6433167