Add generic types to DebugModifiers.
Courtesy of fredg1. This generally reduces unchecked casts from
Object, which safeguards against future misuse of these shared maps by
inserting incorrectly typed objects.

This change also reduces DebugModifers.allModifiers()'s dependence on
mutable global state, which is generally a good thing.

To test this, we use the `modifies` CLI command, which fetches all
modifiers that match some filter string. We select "water level" as
something that was used in an old path (Heavy Rains), which we expect
is unlikely to be used by new content, or conflict with future
modifier names.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20929 b29ace70-8910-0410-8dcc-aa2fc6433167