Consider CoT / Bjorn familiars even if not worn.
This was causing some incorrect maximizer behavior as reported by
phreddrickk. Thanks for the report!

It manifested in two ways:

First: If you only excluded one of the pseudo-slots crownofthrones / buddy-bjorn
but had both corresponding items in your inventory, then Maximizer would decide
that there's one familiar-carrier slot to maximize for, but not actually
remember which one. This would result in potentially swapping out the current
occupant for a different familiar.

Second: if you didn't have a CoT, if the current occupant didn't contribute to
the score, the code would always try to replace it (assuming that Maximizer
decides it wants to equip CoT / Bjorn) -- even if there are no familiars that
improve the score. In that case, the MaximizerSpeculation would leave the
corresponding field unspecified, get confused since there was no familiar to
bjornify, and just equip $item[none] to the back slot.

I don't think there's a good reason to keep this check.

This change also moves MaximizerTest into the maximizer package, since it just
makes more sense, even if we have no plans to rely on package-visible fields or
methods.

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20939 b29ace70-8910-0410-8dcc-aa2fc6433167