Add a RequestTestBase for common setup / helpers.
Right now, this provides some testing the happy path: you specify some
response text for a spy, and we inject that data when you execute
spy.run().

Important: in order for everything to work end-to-end, you need to
provide a spy, not a regular mock. This allows us to exercise as much
real behavior as possible.