Add a placeholder dependent action.
This is effectively a no-op, but we need to perform _some_ action in
each step, and we need at least one step in each job.

This allows us to get around the naming issues where the main branch
expects a check with the name "Run Tests" to complete, while the
matrix strategy modifies the name of each of the subjobs.