Remove Ant build (#66)
* Use grgit for updating the repo from origin/main.

While we're at it, also remove the need for the git CLI for building
from source. We do need this to be run in a git repository with a .git
directory, though...

And, remove the dependency on build.xml and default.properties,
instead setting java.version in gradle.properties.

* Delete build.xml, default.properties, jars.

We haven't yet ported the javadoc task, exe / dmg building, or
whatever experimental coverage viewer there was. We might not need any
of those.

* Also delete build.properties.

I think this is only used in the Ant build.

* Delete the download, toCopy parts of build.gradle.

We no longer want to bundle jars explicitly (as we can / should
instead rely on Gradle to handle dependency management), so there's no
point in downloading them.

* Only count revisions for commits on origin/main.

This should prevent revision drift for those with lots of local
commits.

* Clean up build.gradle further.

- Fix formatting, which doesn't seem to be addressed by spotless for
  some reason.

- Remove svnRevList. We're steering hard into using git.

- Remove redundant dependencies on gitRevList. pruneDist already
  relies on it, so things that rely on pruneDist shouldn't need to
  take an additional dependency.

* Improve gitUpdate functionality.

- Detect when no update is required.

- Pull from the current tracked branch, as opposed to always
  origin/main.

- If an update is required, "stash" the existing changes by creating a
  temp commit, rebase onto that temp commit, and then reset that
  commit.

* Fix gradle formatting to use Eclipse style.

I don't really care about the style as long as it actually works for
nested blocks, which the previous one did not.

* Format gradle files exclusively with greclipse().

I think this is causing check failures.

* Add env variable for overriding the head commit.

This should allow us to pass $GITHUB_SHA when we detect the number of
revisions, rather than always using origin/main (which results in
building the same revision multiple times if PRs are merged in quick
succession).