Migrate JUnit4 tests to JUnit5 (#78)
* Migrate tests to JUnit 5.

The most annoying ones were parameterized tests, since the structure
changed appreciably.

The others were mostly handled automatically by IntelliJ's migration
actions, coupled with some search-and-replace to use assertions
statically.

Note that this is based on the `antless` branch, since I didn't want
to deal with build.gradle merge conflicts shortly thereafter.

* Remove leftover JUnit 4 dependencies.

One test (SkillDatabaseTest) was added after I did the initial
migration.

This outright removes ErrorCollector from DataFileConsistencyTest,
since it no longer exists in JUnit 5. We could continue using it, but
I'd prefer not to if it means we can remove the dependency entirely.

* Remove JUnit 4, vintage engine from build.gradle.

To prevent backsliding in the future.

* Publish test reports.

While there's a comment insisting that it relies on pull_request_target (not
pull_request), the example code for the action suggests using it with
pull_request. Worth a shot, since I have no idea why a test is failing for Mac.

* Convert CustomScriptTest to a parameterized test.

I still have no idea why this is hanging in the MacOS build.

...

Note after the fact: this is no longer hanging in the MacOS build, as of the
migration to parameterized tests. I don't understand, but I'm not complaining.
Could have been a transient issue with the github test runners.