Set username instead of mocking Preferences.
It turns out that mocking static methods of Preferences is slow for
various reasons.

That said, we can set preferences and read per-user defaults if we
just set a username. It creates a default CCS file in test/root, but
that's a small price to pay.

This change also sets the min Java version in the readme to 11, since
I'd missed that the first time around. Again, we don't have to bump
the min version now if we don't want to; I'll just need to type a
little bit more.