Add initial pass of CONTRIBUTING.md. (#87)
* Add initial pass of CONTRIBUTING.md.

There's a lot more room for expansion, but I think this is a good
start.

* Add section on testing, more links.

* Fix spelling error.

archetecture -> architecture.

* Add some links to forum discussions.

We've discussed development practices in a few public threads. The two
I found covered CustomScriptTest and Gradle.

* Wordsmith CONTRIBUTING.md some more.

This rewrites the initial section to make it explicit that this
describes the workflow / process, and not any specific tooling.

This also reworks a couple of paragraphs under Guidelines.