Check for null in UseItemEqnueuePanel.
We apparently need to do this every time we want to access a method of
item. Worth refactoring on a later date so this is less dangerous.

This change also resolves a mismatch between the item name (1950
Vampire Vintner wine) and its presence in inebriety.txt,
modiiers.txt, and statuseffects.txt.

For reasons, we only expanded DataFileConsistencyTest to cover
inebriety.txt (since it best fit the model used by the existing check
of equipment.txt).