Add tests for BasementRequest. (#84)
* Add tests for BasementRequest.

IMPORTANT: this change upgrades our required Java version to 11.
This is the LTS release immediately following 8, so I don't anticipate
as many breakages as the previous bump from 8 to 9.

This isn't complete by any means (we're still missing elemental tests
and HP/MP drain, as well as rewards), but it's a start.

* Bump minimum Java required in README.

* Set username instead of mocking Preferences.

It turns out that mocking static methods of Preferences is slow for
various reasons.

That said, we can set preferences and read per-user defaults if we
just set a username. It creates a default CCS file in test/root, but
that's a small price to pay.

This change also sets the min Java version in the readme to 11, since
I'd missed that the first time around. Again, we don't have to bump
the min version now if we don't want to; I'll just need to type a
little bit more.

* Revert Mockito dependency change.

We're no longer trying to mock static methods, so we should be able to
use mockito-core again.