Coerce first argument of get_player_name to int. (#96)
This caused problems when invoked with a type that's coercible to int,
namely a float. While it was most visible in JS (which only has a
Number type, equivalent to ASH's float), it could also be reproduced
in ASH.

For a more systematic fix, we might want to investigate always
coercing arguments of library functions, rather than doing it on a
one-off basis in each method.

This also fixes issues with CustomScriptTest that were introduced when
migrating it to JUnit5 parameterized tests.