Wrap URL.openConnection.
This controls all current accesses outside of DataUtilities (since
that lives in lib/).

As such, we can then use this in a new JUnit 5 extension that forces
it to return null (thereby generating a NPE for any users, and
probably causing network-dependent tests to fail unless they are
robust against this scenario).