Don't test FightRequest.initializeAfterFight().
This isn't a very interesting aspect of the test, but worse yet, this
interacts with updateCombatData(NOT_IN_A_FIGHT) to generate a request
in a separate thread to log in as the current user.

This is A Bad Thing, in that it probably generates a NPE as a result
of ForbidNetworkAccess, which if uncaught probably causes the JUnit
Platform to be very unhappy with us.