Disable writing settings more directly in tests. (#124)
* Disable writing settings more directly in tests.

We think there may be some settings leaking across tests, which
is... not ideal.

* Disable test parallelism.

This seems like it may be causing some of our problems with flaky
tests.

* Back out previous changes, synchronize instead.

Just experimenting to see if this helps...

* If you scope the reset in the test, it works.  We either need to mock KoLCharacter or we need KoLCharacter.reset in the tests.

* Fixed AscensionPath which was setting property garlandUpgrades instedad of gloverPoints

* Fixed AscensionPath which was setting property garlandUpgrades instedad of gloverPoints

* Let's pass Spotless

* Now with no GarlandUpgrades at all.

* changed Init method to protected

* Fix typo, formatting.

qauntum -> quantum, and also run ./gradlew spotlessApply.

* Remove test session file for fakeUserName.

This shouldn't be created under normal circumstances.

* Add global prefs in test/root.

The most notable setting here is fixedThreadPoolSize=1. That might
help with race conditions...

* Revert "Add global prefs in test/root."

This reverts commit d1c8b5402e5c383516cf5f7cd345f96f8c487fcc, as it
didn't seem to work.

* Add a simple check for test lifetime events.

We're apparently seeing issues saving preferences on Mac. At the very
least, we should make sure that our username is being appropriately
set.

* Change visibility for lifecycle methods.

BeforeAll / AfterAll are not intended to be used with private methods.

(I'm admittedly not sure why these were working before.)

* Wrap a couple more calls to KoLCharacter.reset.

These two were still generating preference files.

* Don't test FightRequest.initializeAfterFight().

This isn't a very interesting aspect of the test, but worse yet, this
interacts with updateCombatData(NOT_IN_A_FIGHT) to generate a request
in a separate thread to log in as the current user.

This is A Bad Thing, in that it probably generates a NPE as a result
of ForbidNetworkAccess, which if uncaught probably causes the JUnit
Platform to be very unhappy with us.

* Track garlandUpgrades AND gloverPoints

Co-authored-by: BadHorseMonkey <58645293+BadHorseMonkey@users.noreply.github.com>
Co-authored-by: Sam Gaus <sam@gaus.co.uk>