Clear Mafia error state after requests. (#135)
* Clear Mafia error state after requests.

In BasementRequestTest, we were clearing the error state before the
request. This is actually backwards, and masked a separate issue where
UseSkillRequest's (tested!) failure was leaking across tests.

* Improve error messages when JS is interrupted.

In general, it's nice to know more than just "Script interrupted."
especially if it was interrupted because Mafia ran into some other
error.