Extend what it means for a build to be "dirty".
In particular, consider it to be dirty if it's not tagged with the
corresponding revision.

This should be more flexible than just checking the HEAD of
origin/main, since it'll continue to work if you build an old
revision.

We also need to make sure that we're tagging revisions before building
jars / binaries for distribution, so they're not all marked as
"dirty".