Extend what it means for a build to be "dirty". (#169)
Consider a build to be dirty if it has any changes not in origin/main.

This should be more flexible than just checking the HEAD of
origin/main, since it'll continue to work if you build an old
revision.

This PR also fixes revision detection for builds of old revisions,
which previously just used "whatever origin/main is at". Now we count
the total number of revisions, minus those that are not present
upstream. (It would be nice if grgit supported git merge-base, but
this works well enough.)