Use cloudfront for images, take two. (#314)
KoL is currently serving a mix of cloudfront / S3 traffic, so we should be
strict about requesting from cloudfront if we originate a request, but lenient
about accepting responses that contain non-cloudfront URLs.

Further, downloadImage() was constrained such that it would look for the last
occurrence of .com, which naturally stopped matching once we moved to
cloudfront.net URLs.

Also also, there were a bunch of regexes that expected image URLs to contain
"images.kingdomofloathing.com", which was true under S3 but is no longer true
with Cloudfront. Many of these are underspecified out of laziness, but I can fix
that if needed.