unused urlString parameter this was there because I initially thought that any request not in adventure.php or fight.php wouldn't affect predictions
It's worth noting that while that assertion ended up not being true,
we'll most likely end up needing to use it again later.
This is because relying on KoLAdventure.lastLocationName is error prone,
because we only want to list visits to adventure.php locations,
and KoLAdventure.lastLocationName currently includes
locations such as cellar.php.

(I guess we could also try to spot them at the start of the method...)