Add some tests to the maximizer (#382)
The maximizer is very big and complicated, and a good target for additional tests before refactoring (or in general).

I refactored the tests a bit to prefer using names to ids; I feel it reads better.

I added tests for max and min (inside modifiers, or starting for max), stat restrictions and clownosity. The documentation says that max and min at the start of an expression should apply to the entire expression, but I tried this and I don't think it works that way for min. It looks like "min" fails the maximization if the current character items / effects don't meet the minimum.