Improve variable reference labeling.
`Variable 'null' ... ` is not a very useful message. Instead, we
should quote a small amount of text from the input script.

There's likely value in better handling bad locations (e.g. by
throwing an exception), but I posit that a bad location is indicative
of a bigger problem in parsing.