Recognize bad record indexing without throwing exception (#806)
* Catch exceptions in variable initialization.

This also updates AshRuntime to print all error messages, not just the
first one.

* CompositeReference might have BadType indicies

* Record type returns BadType on bad index type, rather than throw internal error

* add tests

* Improve variable reference labeling.

`Variable 'null' ... ` is not a very useful message. Instead, we
should quote a small amount of text from the input script.

There's likely value in better handling bad locations (e.g. by
throwing an exception), but I posit that a bad location is indicative
of a bigger problem in parsing.

Co-authored-by: Veracity0 <83725298+Veracity0@users.noreply.github.com>