Only warn for the Kung Fu Fighting intrinsic.
The non-intrinsic effect can be acquired via pocket wishes / hookah /
other sources, and does not go away if entering a combat with equipped
weapons / offhands.

This change makes a field and a method use package visibility, which
is more restrictive than `public` -- I don't envision any use cases
for this outside of testing.