refactor: split ModifierDatabase from Modifiers.
Roughly, ModifierDatabase tries to handle those actions which are not
character-specific: reading from the database in modifiers.txt, writing
to it, parsing it, parsing KoL descriptions.

There is one exception: `getNumericModifier` for FamiliarData sets the
user's current familiar on the Modifiers object as a static field. There
are a few of these fields. This is used in `Expression` when calculating
familiar-specific modifiers. I feel like we shouldn't do it this way.

One function that was only used in `BasementRequest` was moved there. I
felt based on how it worked it wasn't usable in general.