refactor: split ModifierDatabase from Modifiers (#1467)
* refactor: split ModifierDatabase from Modifiers.

Roughly, ModifierDatabase tries to handle those actions which are not
character-specific: reading from the database in modifiers.txt, writing
to it, parsing it, parsing KoL descriptions.

There is one exception: `getNumericModifier` for FamiliarData sets the
user's current familiar on the Modifiers object as a static field. There
are a few of these fields. This is used in `Expression` when calculating
familiar-specific modifiers. I feel like we shouldn't do it this way.

One function that was only used in `BasementRequest` was moved there. I
felt based on how it worked it wasn't usable in general.

* chore: spotless

* chore: replace fully qualified name with import

* fix: initialise modifier database in modifiers class

Co-authored-by: Jamie Adams <82782908+jaadams5@users.noreply.github.com>