Explicitly check for null-ness of SESSION_HOLDER.
SESSION_HOLDER is set to a non-null value if we managed to acquire an
exclusive lock. Further, to maintain this invariant, when we release
the lock, reset SESSION_HOLDER to null.

SESSION_FILE isn't as good of a fit, as we will set it to a non-null
value when we open the file for reading before trying to acquire an
exclusive lock. It's currently not actively harmful because we'll just
run into a NPE when we call SESSION_HOLDER.release(), but that's still
not ideal.