Release held session file lock before acquiring a new one (#1479)
* Add message for file lock held by another process.

Currently, this just silently fails. That's not great.

* Release held file lock before acquiring a new one.

Previously, we tied this to LogoutManager. However, given that a
session can be ended without going through LogoutMAnager
(e.g. sleeping through rollover, session timeout, creating a new
session) and LoginManager can be invoked again, this breaks an
existing use case.

* Explicitly check for null-ness of SESSION_HOLDER.

SESSION_HOLDER is set to a non-null value if we managed to acquire an
exclusive lock. Further, to maintain this invariant, when we release
the lock, reset SESSION_HOLDER to null.

SESSION_FILE isn't as good of a fit, as we will set it to a non-null
value when we open the file for reading before trying to acquire an
exclusive lock. It's currently not actively harmful because we'll just
run into a NPE when we call SESSION_HOLDER.release(), but that's still
not ideal.

---------

Co-authored-by: Veracity0 <83725298+Veracity0@users.noreply.github.com>
Co-authored-by: Chris Midgley <chris.midgley@dunecomputers.co.uk>