Move some function calls to the Swing thread.
Various listener updates can cause Swing changes in rendering etc, and
since the Swing thread has a dependency on Preferences.lock, we want to
avoid a deadlock.

WARNING: This may resolve in a noticeable slowdown / visual hiccups /
stalls in the GUI, as we're moving noticeably more processing into the
Swing thread as a result.

This also updates DeadlockDetector to print out the rest of the lines of
the stack trace, since 8 lines is often not enough for debugging these deep
call stacks.