Ignore failed rests (#1970)
* Ignore failed rests

If you try to rest when it is pointless to do so (full HP, full MP, no Beaten Up), the game won't let you do it, and will display a message "You don't need to rest right now." instead. This does not decrement the amount of free rests you can use, and this does not subtract an Adventure if you are out of free rests. However, KoLmafia thinks that the rest did happen, and increases timesRested accordingly. This leads to the undercounting of the amount of free rests left (Mafia fixes the data if it sees that a free rest is actually available, but it thinks it is not; however, the information that there is more than one free rest would be lost). The fix is to not increment timesRested in the case of a failed rest.

* Added a test processor

Forgot to commit initially.

* Try to appease Gradle gods

Because Gradle does not want to work on my machine.

* Try to appease Gradle gods, attempt 2

* Fix the test

Forgot to remove a line from the copied test.

* Remove extra symbols in comment

* Add a test on a successful rest

Also move both into a nested class.