Fix skill type for Ravenous Pounce.
This is a combat skill with a passive effect.

In particular, this was previously correctly categorized, since Stomach
Capacity was not a modifier until recently.

It probably makes sense to add a test to guard against adding a non-passive
skill that has an associated modifier, although with new skills, this is
less likely to be an issue (since modifier tracking simply won't work if
implemented as the wrong type).