feat: add boolean modifiers for extra pickpocket attempts and negative status resists (#2170)
Neither stack with other equipment that offer the same modifiers: no
matter how many "75% chance to resist negative statuses" you have, it's
still 75%. Similarly, having both "extra pickpocket" items still only
gives you one extra pickpocket.

Add the negative status resist because there are enough of them that we
may as well track it. Add the pickpockets because this is absurdly good
in certain fringe scenarios, and it's easier to stay aware of it if
Mafia tracks it explicitly.