feat: allow trailing undefined values to be passed to ash functions
Javascript allows callers to pass additional arguments to a function,
even if they were not declared in the function signature, or omit
parameters from a function call, which will make them have an
`undefined` value.
To accommodate this, Typescript allows parameters to be marked as
optional, which allows the caller to either omit a trailing parameter,
or explicitly pass `undefined` for that parameter.

We have several RuntimeLibrary functions where the last parameter is
optional, because there is an overladed version where all other
parameter are identical. While omitting parameters works in this case,
explicitly passing `undefined` does not.

To make this more convenient to use, and more compatible with
Javascript/Typescript convention, strip `undefined` parameters from the
end of the arguments array. Since undefined values are (so far) not at
all supported by either scripting environment, this should not pose any
incompatiblity issues.